# openapi.yaml
openapi: 3.1.1
info:
  title: Open-Spot API
  description: |
    개인 장소 기록 및 관리 서비스 API

    ## 주요 기능
    - Google OAuth2 인증 (Bearer Token + HttpOnly Cookie)
    - 개인 장소 기록 관리 (평점, 리뷰, 태그)
    - 장소 그룹 관리
    - 푸시 알림 관리

    ## 인증 방식
    - **Access Token**: Authorization 헤더로 전달 (Bearer 방식)
    - **Refresh Token**: HttpOnly Cookie로 전달 (XSS 방어)

    ## MSA 아키텍처
    - **Gateway (8080)**: 모든 API의 진입점, JWT 검증 수행
    - **Auth Service (8081)**: Google OAuth2 로그인, JWT 발급
    - **Location Service (8082)**: 장소 관리
    - **Notification Service (8083)**: 알림 관리

    **중요**: 프론트엔드는 반드시 Gateway(8080)를 통해 API를 호출해야 합니다.
  version: 1.0.0
  contact:
    name: Open-Spot Team
servers:
  - url: http://localhost:8080
    description: Gateway (모든 API 호출은 이 URL 사용)
  - url: http://localhost:8081
    description: Auth Service (직접 호출 금지, Gateway 경유 필수)
  - url: http://localhost:8082
    description: Location Service (직접 호출 금지, Gateway 경유 필수)
  - url: http://localhost:8083
    description: Notification Service (직접 호출 금지, Gateway 경유 필수)
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/locations/{locationId}:
    get:
      tags:
        - Locations
      summary: 장소 상세 조회
      description: 특정 장소의 상세 정보를 조회합니다.
      operationId: getLocationById
      parameters:
        - $ref: '#/components/parameters/LocationIdParam'
      responses:
        '200':
          description: 장소 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LocationResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Locations
      summary: 장소 기본 정보 수정
      description: 장소의 기본 정보를 수정합니다.
      operationId: updateLocation
      parameters:
        - $ref: '#/components/parameters/LocationIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLocationRequest'
      responses:
        '200':
          description: 장소 수정 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LocationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Locations
      summary: 장소 비활성화
      description: 장소를 비활성화합니다 (논리적 삭제).
      operationId: deleteLocation
      parameters:
        - $ref: '#/components/parameters/LocationIdParam'
      responses:
        '200':
          description: 장소 비활성화 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          locationId:
                            type: string
                            format: uuid
                          message:
                            type: string
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/locations/groups/{groupId}:
    get:
      tags:
        - Locations
      summary: 그룹별 장소 목록
      description: 특정 그룹에 속한 내 장소 목록을 조회합니다.
      operationId: getLocationsByGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            nullable: true
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 그룹별 장소 목록 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LocationSummaryPageResponse'

    put:
      tags:
        - Location Groups
      summary: 그룹 수정
      description: 그룹의 기본 정보를 수정합니다.
      operationId: updateLocationGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLocationGroupRequest'
      responses:
        '200':
          description: 그룹 수정 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LocationGroupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Location Groups
      summary: 그룹 삭제
      description: 그룹을 삭제합니다.
      operationId: deleteLocationGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 그룹 삭제 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          groupId:
                            type: string
                            format: uuid
                          message:
                            type: string
        '404':
          $ref: '#/components/responses/NotFound'

  # ========================================
  # Location Groups APIs
  # ========================================
  /api/v1/locations/groups:
    get:
      tags:
        - Location Groups
      summary: 그룹 목록 조회
      description: 사용자의 모든 그룹 목록을 조회합니다.
      operationId: getLocationGroups
      responses:
        '200':
          description: 그룹 목록 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/LocationGroupResponse'

    post:
      tags:
        - Location Groups
      summary: 그룹 생성
      description: 새로운 장소 그룹을 생성합니다.
      operationId: createLocationGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationGroupRequest'
      responses:
        '201':
          description: 그룹 생성 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LocationGroupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/locations/groups/reorder:
    put:
      tags:
        - Location Groups
      summary: 그룹 순서 변경
      description: 드래그 앤 드롭으로 그룹 순서를 변경합니다.
      operationId: reorderLocationGroups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderLocationGroupsRequest'
      responses:
        '200':
          description: 그룹 순서 변경 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/LocationGroupResponse'

  # ========================================
  # Notifications APIs
  # ========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Access Token을 Authorization 헤더로 전달합니다.

        **사용법**: `Authorization: Bearer <access-token>`

        **토큰 획득**:
        1. Google 로그인 성공 시 URL 쿼리 파라미터로 전달
        2. 토큰 갱신 API 응답 본문에서 획득

        **인증 플로우 (MSA 표준)**:
        1. 클라이언트 → Gateway: `Authorization: Bearer <token>` 헤더 포함
        2. Gateway: JWT 검증 (HS256 알고리즘)
        3. Gateway: 검증 성공 시 `X-User-Id` 헤더 추가
        4. 내부 서비스: `X-User-Id` 헤더로 사용자 식별 (JWT 검증 불필요)
  parameters:
...
    LocationIdParam:
      name: locationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        data:
          type: object
          nullable: true
        error:
          type: object
          nullable: true
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    PageInfo:
      type: object
      required:
        - number
        - size
        - totalElements
        - totalPages
        - first
        - last
      properties:
        number:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean

    # ========================================
    # Authentication Schemas
    # ========================================
...
    CategoryInfo:
      type: object
      required:
        - id
        - code
        - displayName
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        displayName:
          type: string
        icon:
          type: string
          nullable: true
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          nullable: true

    # ========================================
    # Location Schemas
    # ========================================
    CoordinatesResponse:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
...
    UpdateLocationRequest:
      type: object
      required:
        - name
        - categoryId
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 1000
          nullable: true
        address:
          type: string
          maxLength: 200
          nullable: true
        categoryId:
          type: string
          format: uuid
        iconUrl:
          type: string
          format: uri
          maxLength: 500
          nullable: true
...
    LocationResponse:
      type: object
      required:
        - id
        - userId
        - name
        - categoryId
        - category
        - coordinates
        - isActive
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        categoryId:
          type: string
          format: uuid
        category:
          $ref: '#/components/schemas/CategoryInfo'
        coordinates:
          $ref: '#/components/schemas/CoordinatesResponse'
        iconUrl:
          type: string
          format: uri
          nullable: true
        personalRating:
          type: integer
          nullable: true
        personalReview:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        groupId:
          type: string
          format: uuid
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        distance:
          type: number
          format: double
          nullable: true
    LocationSummaryResponse:
      type: object
      required:
        - id
        - name
        - categoryId
        - category
        - coordinates
        - tags
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        categoryId:
          type: string
          format: uuid
        category:
          $ref: '#/components/schemas/CategoryInfo'
        coordinates:
          $ref: '#/components/schemas/CoordinatesResponse'
        personalRating:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            type: string
        distance:
          type: number
          format: double
          nullable: true
    LocationSummaryPageResponse:
      type: object
      required:
        - content
        - page
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/LocationSummaryResponse'
        page:
          $ref: '#/components/schemas/PageInfo'

    # ========================================
    # Location Group Schemas
    # ========================================
    CreateLocationGroupRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
          nullable: true
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          nullable: true
        icon:
          type: string
          maxLength: 50
          nullable: true
    UpdateLocationGroupRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
          nullable: true
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          nullable: true
        icon:
          type: string
          maxLength: 50
          nullable: true
    ReorderLocationGroupsRequest:
      type: object
      required:
        - groupOrders
      properties:
        groupOrders:
          type: array
          items:
            $ref: '#/components/schemas/GroupOrderItem'
    GroupOrderItem:
      type: object
      required:
        - groupId
        - order
      properties:
        groupId:
          type: string
          format: uuid
        order:
          type: integer
          minimum: 0
    LocationGroupResponse:
      type: object
      required:
        - id
        - userId
        - name
        - order
        - isShared
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        order:
          type: integer
        isShared:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ========================================
    # Notification Schemas
    # ========================================
...